Arreglo de errores

+++ SegundoParcial/Entidades/Vehiculo.cs
@@ -13,7 +13,7 @@ namespace SegundoParcial.Entidades
         [Key]
         public int VehiculoId { get; set; }
         public string Descripcion { get; set; }
-        public int MantenimientoTotal { get; set; }
+        public decimal MantenimientoTotal { get; set; }
 
         public Vehiculo(){ }
